% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_fun.R
\name{death_fun}
\alias{death_fun}
\title{Function that performs a death step}
\usage{
death_fun(k, k_new, a, sig2, x, X, kmax)
}
\arguments{
\item{k}{:int}

\item{k_new}{:int}

\item{a}{: vec}

\item{sig2}{:num}

\item{x}{:vec}

\item{X}{:matrix}

\item{kmax}{:int}
}
\value{
list with updated a and k values
}
\description{
Function that performs a death step
}
\examples{
## Will add soon
}
