% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mixture.R
\name{get_lengths}
\alias{get_lengths}
\title{Get Lengths of Jump Matrix Columns Across Chains}
\usage{
get_lengths(outputs, n_chain)
}
\arguments{
\item{outputs}{List. A list containing the \code{jump} matrices for all chains.
Typically, this is the output from an RJMCMC process.}

\item{n_chain}{Integer. The number of chains to process.}
}
\value{
A named vector representing the sorted frequency table of column lengths
across all chains.
}
\description{
This function calculates the number of columns (lengths) in the \code{jump} matrix
for each chain within the \code{outputs} object and compiles the results into a data frame.
It also creates a sorted frequency table of the unique lengths in descending order.
}
\details{
The function iterates through each chain in the range \code{1:n_chain}. For each chain,
it calculates the number of columns in each \code{jump} matrix stored in \code{outputs$jump}.
These lengths are compiled into a data frame with two columns:
\itemize{
\item \code{chain}: The chain index.
\item \code{length_n}: The number of columns in the \code{jump} matrix for that chain.
}

Finally, it creates a frequency table of the \code{length_n} values and sorts it
in descending order, allowing the user to see the most common lengths of \code{jump} matrices.
}
