% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmc_full.R
\name{get_output_full}
\alias{get_output_full}
\title{Get Output from Reversible Jump Monte Carlo Monte Carlo (RJMC) Simulation}
\usage{
get_output_full(model, data_list, settings, update_ind, par)
}
\arguments{
\item{model}{A list representing the model to be used in the RJMC simulation. This typically includes information
such as model parameters, parameter names, and other model-specific settings required for the simulation.}

\item{data_list}{A list containing the data used by the model. This could include observed data, prior distributions,
or other inputs necessary for the RJMC simulation.}

\item{settings}{A list of settings for the RJMC simulation. The settings must include:
\itemize{
\item \code{numberChainRuns}: The number of parallel chains to run.
\item \code{runParallel}: Logical, whether to run the chains in parallel.
\item \code{numberFittedPar}: The number of parameters to fit in the model.
\item \code{numberCores}: The number of CPU cores to use for parallel execution (if \code{runParallel = TRUE}).
\item Other relevant settings that influence the RJMC process, such as iteration count, burn-in period, etc.
}}

\item{update_ind}{Logical flag indicating whether to update the model parameters. Typically, this is used to specify
whether parameters should be updated during the simulation.}

\item{par}{A list of parameters for each chain. Each element of the list corresponds to the parameter values for one
chain. If provided, these values are used for the chains; otherwise, the simulation will use default
parameter values.}
}
\value{
A list containing the processed output of the RJMC simulation, which includes:
\itemize{
\item \code{mcmc}: An \code{mcmc.list} object containing the posterior samples from each chain.
\item \code{lpost}: A data frame of log-posterior values for each sample and chain.
\item \code{temp}: A data frame of temperatures for each chain.
\item \code{acc}: A data frame of acceptance rates for each chain.
\item \code{outRJpar}: A list of parameter values for each chain during the simulation.
}
}
\description{
This function performs the core steps of running multiple chains of the Reversible Jump Monte Carlo (RJMC)
simulation and processes the results. It either runs the chains in parallel or sequentially, collects the output
from each chain, and organizes it into a list format. The output includes posterior samples, log-posterior values,
temperatures, acceptance rates, and other diagnostic information for each chain.
}
\details{
This function is designed to handle both parallel and sequential execution of RJMC chains. If \code{runParallel} is \code{TRUE},
the chains are executed in parallel using the \code{mclapply} function. Otherwise, the chains are executed sequentially
in a loop. The output is processed and reshaped into appropriate data frames for easy analysis.
}
