% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_fun.R
\name{get_r_fun}
\alias{get_r_fun}
\title{Function that calculates the r_(i,j) probability in the acceptance ratio}
\usage{
get_r_fun(k_int, k_prop, kmax)
}
\arguments{
\item{k_int}{: integer in \link{0,kmax}, current k value}

\item{k_prop}{: integer in \link{0,kmax}, either k_int - 1 or k_int + 1, proposed k value}

\item{kmax}{: integer greater than 0, max ar(k) process that can be checked}
}
\value{
probability of either birth or death
}
\description{
Function that calculates the r_(i,j) probability in the acceptance ratio
}
\examples{
will add soon
}
