% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_fun.R
\name{k_fun}
\alias{k_fun}
\title{Determines a proposal for new k}
\usage{
k_fun(k, kmax)
}
\arguments{
\item{k}{:int}

\item{kmax}{:int}
}
\value{
an integer that is the proposal k
}
\description{
Determines a proposal for new k
}
\examples{
will add soon
}
