% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike_fun.R
\name{loglike_fun}
\alias{loglike_fun}
\title{Calculates the log-likelihood}
\usage{
loglike_fun(kmax, k, sig2, a, x, X, maxT)
}
\arguments{
\item{kmax}{:int}

\item{k}{:int}

\item{sig2}{:num}

\item{a}{: vec}

\item{x}{:vec}

\item{X}{:matrix}

\item{maxT}{: int}
}
\value{
the log likelihood
}
\description{
Calculates the log-likelihood
}
\examples{
## will add soon
}
