% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc_nested.R
\name{rjmcmc_nested}
\alias{rjmcmc_nested}
\title{Runs the rjmcmc algorithm}
\usage{
rjmcmc_nested(iter, k, sig2, x, kmax, alpha_p = 2, beta_p = 1)
}
\arguments{
\item{iter}{}

\item{k}{}

\item{sig2}{}

\item{x}{}

\item{kmax}{}

\item{alpha_p}{}

\item{beta_p}{}
}
\value{
matrix with parameters a, k, and sig2
}
\description{
Runs the rjmcmc algorithm
}
\examples{
will add soon
}
