% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_fun.R
\name{within_fun}
\alias{within_fun}
\title{function to perform a within model move}
\usage{
within_fun(k, x, X, sig2)
}
\arguments{
\item{k}{: int}

\item{x}{:vec}

\item{X}{: matrix}

\item{sig2}{: num}
}
\value{
updated ar coefficients a
}
\description{
function to perform a within model move
}
\examples{
will add soon
}
