% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_matrix_creation.R
\name{X_matrix_creation}
\alias{X_matrix_creation}
\title{Function creates lagged x matrix}
\usage{
X_matrix_creation(x, maxT, kmax)
}
\arguments{
\item{x}{: time series: column vector (Tmax x 1)}

\item{maxT}{: integer value > 0: length of time series x}

\item{kmax}{: integer value > 1: max value of k in an AR(k) can potentially move to}
}
\value{
matrix of lagged x time series
}
\description{
Function creates lagged x matrix
}
\examples{
x = c(4.8407828, 3.1495432, 1.0391882, -0.5670270, -1.2837761, -0.4528256,
 -1.1621543, -1.8725529, -2.7431505, -3.6796526)        
maxT = 10        
kmax = 30         
        
        
}
