% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_fun.R
\name{birth_fun}
\alias{birth_fun}
\title{Function that performs a birth step}
\usage{
birth_fun(k, k_new, a, sig2, x, X, kmax)
}
\arguments{
\item{k}{: int}

\item{k_new}{: int}

\item{a}{vec}

\item{sig2}{: number}

\item{x}{:vector}

\item{X}{: matrix}

\item{kmax}{:int}
}
\value{
a list with updated a and k values
}
\description{
Function that performs a birth step
}
\examples{
Will add soon
}
