% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc_nested.R
\name{rjmcmc_nested}
\alias{rjmcmc_nested}
\title{Runs the rjmcmc algorithm}
\usage{
rjmcmc_nested(iter, k, sig2, x, kmax, alpha_p = 2, beta_p = 1)
}
\arguments{
\item{iter}{:int}

\item{k}{:int}

\item{sig2}{:num}

\item{x}{: vec}

\item{kmax}{: int}

\item{alpha_p}{:int}

\item{beta_p}{:int}
}
\value{
matrix with parameters a, k, and sig2
}
\description{
Runs the rjmcmc algorithm
}
\examples{
## will add soon
}
\references{
\insertRef{brooks_efficient_2003}{rjmc}

\insertRef{green_reversible_1995}{rjmc}

\insertRef{robert_monte_2004}{rjmc}
}
