% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc_nested.R
\name{rjmcmc_nested}
\alias{rjmcmc_nested}
\title{Runs the rjmcmc algorithm}
\usage{
rjmcmc_nested(iter, k, sig2, x, kmax, alpha_p = 2, beta_p = 1)
}
\arguments{
\item{iter}{:int number of iterations for a single simulation}

\item{k}{:int initial value for the k parameter}

\item{sig2}{:float initial value for sig2 parameter}

\item{x}{: vector (Tmax x 1) of data from the time series}

\item{kmax}{: integer value > 1: max value of k can potentially move to}

\item{alpha_p}{:float shape parameter for  for inverse gamma prior on sig2}

\item{beta_p}{:float scale parameter for inverse gamma prior on sig2}
}
\value{
matrix with parameters a, k, and sig2
}
\description{
Runs the rjmcmc algorithm
}
\examples{
## will add soon
}
\references{
\insertRef{brooks_efficient_2003}{rjmc}

\insertRef{green_reversible_1995}{rjmc}

\insertRef{robert_monte_2004}{rjmc}
}
