% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_matrix_creation.R
\name{X_matrix_creation}
\alias{X_matrix_creation}
\title{Function creates lagged x matrix}
\usage{
X_matrix_creation(x, maxT, kmax)
}
\arguments{
\item{x}{: time series: column vector (Tmax x 1)}

\item{maxT}{: integer value > 0: length of time series x}

\item{kmax}{: integer value > 1: max value of k in an AR(k) can potentially move to}
}
\value{
matrix of lagged x time series
}
\description{
Function creates lagged x matrix
}
\examples{
Example to come
}
